requirements.txt
keep-alive-ping
flask
dotenv
Flask==2.1.2
Werkzeug==2.1.2
pytz==2023.3
requests
apscheduler
python-telegram-bot
google-generativeai 
google-genai
firebase_admin
jsonpickle
pyrate-limiter==2.3.3
cachetools 
pillow
aiofiles
psutil
speedtest-cli
telegraph[aio]
search_engine_parser==0.6.8
bing-image-downloader==1.1.2
wikipedia
youtube_search
yt_dlp
